+++ internal/extconfig.go
@@ -1,0 +1,17 @@
+package internal
+
+import (
+	"fmt"
+
+	"github.com/masamerc/sevp/internal/extconfig"
+)
+
+// GetExternalConfigSelector returns the appropriate Selector implementation based on the selectorName.
+func GetExternalConfigSelector(selectorName string) (Selector, error) {
+	switch selectorName {
+	case "aws":
+		return extconfig.NewAWSProfileSelector(), nil
+	default:
+		return nil, fmt.Errorf("the external config provider is not supported for selector %s", selectorName)
+	}
+}
